Minor project restructure
