+++ src/calc/calc.h
@@ -4,25 +4,6 @@
 #include "../includes/common.h"
 #include "../stack/stack.h"
 
-struct op_type {
-    char op[8];
-    short precedence;
-    short associativity;
-    long double (*eval)(long double a, long double b);
-};
-
-enum {
-    RIGHT,
-    LEFT,
-    NONE
-};
-
-struct Tokens *convert_to_rpn(struct Tokens *expression);
-
-bool is_fun(size_t op);
-bool is_unary(size_t op);
-bool is_op(size_t op);
-
 long double calculate(struct Tokens *expr, long double x);
 
 long double eval_add(long double a, long double b);
@@ -42,5 +23,4 @@ long double eval_arctan(long double a, long double b);
 long double eval_sqrt(long double a, long double b);
 long double eval_ln(long double a, long double b);
 long double eval_log(long double a, long double b);
-
 #endif  // PARSER_H_
