+++ src/test/test_calc.h
@@ -1,10 +1,11 @@
 #ifndef TEST_CALC_H_
 #define TEST_CALC_H_
 
+#include "../calc/calc.h"
 #include "../credit_calc/credit_calc.h"
 #include "../deposit_calc/deposit_calc.h"
-#include "../parser/lexer.h"
-#include "../parser/parser.h"
+#include "../lexer/lexer.h"
+#include "../shunting_yard/shunting.h"
 #include <check.h>
 #include <stdio.h>
 #include <stdlib.h>
